Extract shared addEmailAddress into Create/Update base builders
Introduce AbstractCreateRequestBuilder and AbstractUpdateRequestBuilder
holding a single addEmailAddress() definition per request variant, so the
People/Companies builders inherit it instead of each repeating the method.
This removes at its root the cross-file duplication SonarCloud flagged on
new code (the four parallel addEmailAddress bodies), without any
scanner-side exclusion. Behavior is unchanged: addEmailAddressEntry() in
AbstractRequestBuilder still trims the stored address.

Also drops the earlier sonar.cpd.exclusions attempt, which SonarCloud's
automatic analysis mode ignores.
