--- sonar-project.properties
@@ -1,9 +1,0 @@
-# SonarCloud project configuration.
-#
-# The request builders under src/RequestBuilder are intentionally parallel:
-# People/Companies × Create/Update share the same shape by design (each maps a
-# fluent builder onto a request DTO). Copy-paste detection therefore flags the
-# structurally identical blocks between them as duplication, even though the
-# behavior differs per entity. Exclude these files from CPD so the metric
-# reflects genuine, unintended duplication rather than this deliberate symmetry.
-sonar.cpd.exclusions=src/RequestBuilder/**/*.php
