+++ src/RequestBuilder/Companies/CreateRequestBuilder.php
@@ -25,7 +25,7 @@ use Netresearch\Sdk\CentralStation\Request\Positions;
 use Netresearch\Sdk\CentralStation\Request\RequestInterface;
 use Netresearch\Sdk\CentralStation\Request\Tag;
 use Netresearch\Sdk\CentralStation\Request\Tags;
-use Netresearch\Sdk\CentralStation\RequestBuilder\AbstractRequestBuilder;
+use Netresearch\Sdk\CentralStation\RequestBuilder\AbstractCreateRequestBuilder;
 use Netresearch\Sdk\CentralStation\Validator\Companies\CreateValidator;
 
 use function in_array;
@@ -39,7 +39,7 @@ use function in_array;
  *
  * @api
  */
-class CreateRequestBuilder extends AbstractRequestBuilder
+class CreateRequestBuilder extends AbstractCreateRequestBuilder
 {
     /**
      * Sets the company's data.
@@ -134,21 +134,6 @@ class CreateRequestBuilder extends AbstractRequestBuilder
             'type'        => $type,
             'phoneNumber' => $phoneNumber,
         ];
-
-        return $this;
-    }
-
-    /**
-     * Adds an email address attribute.
-     *
-     * @param string $type         The type of the email address (use one of Constants::CONTACT_DETAILS_TYPE)
-     * @param string $emailAddress The email address
-     *
-     * @return CreateRequestBuilder
-     */
-    public function addEmailAddress(string $type, string $emailAddress): CreateRequestBuilder
-    {
-        $this->addEmailAddressEntry(null, $type, $emailAddress);
 
         return $this;
     }
