+++ src/RequestBuilder/Companies/UpdateRequestBuilder.php
@@ -23,7 +23,7 @@ use Netresearch\Sdk\CentralStation\Request\Positions;
 use Netresearch\Sdk\CentralStation\Request\RequestInterface;
 use Netresearch\Sdk\CentralStation\Request\Tag;
 use Netresearch\Sdk\CentralStation\Request\Tags;
-use Netresearch\Sdk\CentralStation\RequestBuilder\AbstractRequestBuilder;
+use Netresearch\Sdk\CentralStation\RequestBuilder\AbstractUpdateRequestBuilder;
 use Netresearch\Sdk\CentralStation\Validator\Companies\UpdateValidator;
 
 use function in_array;
@@ -37,7 +37,7 @@ use function in_array;
  *
  * @api
  */
-class UpdateRequestBuilder extends AbstractRequestBuilder
+class UpdateRequestBuilder extends AbstractUpdateRequestBuilder
 {
     /**
      * Sets the company's data.
@@ -137,25 +137,6 @@ class UpdateRequestBuilder extends AbstractRequestBuilder
             'type'        => $type,
             'phoneNumber' => $phoneNumber,
         ];
-
-        return $this;
-    }
-
-    /**
-     * Adds an email address attribute.
-     *
-     * @param int|null    $id           The ID of the record to update
-     * @param string|null $type         The type of the email address (use one of Constants::CONTACT_DETAILS_TYPE)
-     * @param string|null $emailAddress The email address
-     *
-     * @return UpdateRequestBuilder
-     */
-    public function addEmailAddress(
-        ?int $id = null,
-        ?string $type = null,
-        ?string $emailAddress = null,
-    ): UpdateRequestBuilder {
-        $this->addEmailAddressEntry($id, $type, $emailAddress);
 
         return $this;
     }
